# webclient/api-reference/api-spec.yaml
openapi: "3.0.3"
info:
  title: "DiagramAsText API"
  description: |
    API to generate diagrams programmatically.
  version: "0.0.6"
  contact:
    email: contact@diagramastext.dev
    name: to access, and to discuss usage conditions and special requests
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "{{ .Env.SERVER }}"
security:
  - BearerAuth: [ ]
tags:
  - name: "Generate Diagram"
    description: |
      The API methods to generate diagrams using description in English. 
      
      <strong>Note</strong>: <strong><i>not idempotent</i></strong>, the same prompt may yield different results.
  - name: "Operations"
    description: The API methods to access API usage quotas and configurations.
paths:
  # diagram generation methods
  /generate/c4:
    post:
      tags:
        - "Generate Diagram"
      summary: "Generates C4 Containers diagram"
      description: "The method generates C4 Container diagram as SVG."
      requestBody:
        description: "Input prompt in plain English"
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/RequestGenerateDiagram"
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ResponseDiagramSVG"
        "400":
          description: Invalid request format
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Usage quota exceeded
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Invalid input prompt
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Throttling quota exceeded
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
  # operations
  /quotas:
    get:
      tags:
        - "Operations"
      summary: "Fetches usage quotas"
      description: "The method to fetch current usage status and the quotas limits left"
      responses:
          "200":
            description: OK
            content:
              "application/json":
                  schema:
                    $ref: "#/components/schemas/Quotas"
          "401":
            description: Unauthorized
            content:
              "application/json":
                schema:
                  $ref: "#/components/schemas/Error"
          "429":
            description: Throttling quota exceeded
            content:
              "application/json":
                schema:
                  $ref: "#/components/schemas/Error"
          "500":
            description: Server error
            content:
              "application/json":
                schema:
                  $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "The diagramastext.dev API key"
  schemas:
    RequestGenerateDiagram:
      example: { "prompt": "C4 diagram of a Go web server reading from external Postgres database over TCP" }
      type: object
      required:
        - "prompt"
      additionalProperties: false
      properties:
        prompt:
          description: "Diagram description in plain English."
          type: "string"
          minLength: 3
    ResponseDiagramSVG:
      example: { "svg": "\u003c?xml version=\"1.0\" encoding=\"us-ascii\" standalone=\"no\"?\u003e\u003csvg xmlns=\"http://www.w3.org/2000/svg\" xmlns:xlink=\"http://www.w3.org/1999/xlink\" contentStyleType=\"text/css\" height=\"237px\" preserveAspectRatio=\"none\" style=\"width:438px;height:237px;background:#FFFFFF;\" version=\"1.1\" viewBox=\"0 0 438 237\" width=\"438px\" zoomAndPan=\"magnify\"\u003e\u003cdefs/\u003e\u003cg\u003e\u003c!--entity 0--\u003e\u003cg id=\"elem_0\"\u003e\u003crect fill=\"#438DD5\" height=\"117.7813\" rx=\"2.5\" ry=\"2.5\" style=\"stroke:#3C7FC0;stroke-width:0.5;\" width=\"189\" x=\"7\" y=\"7\"/\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"16\" font-weight=\"bold\" lengthAdjust=\"spacing\" textLength=\"40\" x=\"49\" y=\"31.8516\"\u003eWeb\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"16\" font-weight=\"bold\" lengthAdjust=\"spacing\" textLength=\"6\" x=\"89\" y=\"31.8516\"\u003e\u0026#160;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"16\" font-weight=\"bold\" lengthAdjust=\"spacing\" textLength=\"59\" x=\"95\" y=\"31.8516\"\u003eServer\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"12\" font-style=\"italic\" lengthAdjust=\"spacing\" textLength=\"26\" x=\"88.5\" y=\"46.7637\"\u003e[Go]\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"99.5\" y=\"62.5889\"\u003e\u0026#160;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"43\" x=\"28.5\" y=\"78.8857\"\u003eReads\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"71.5\" y=\"78.8857\"\u003e\u0026#160;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"35\" x=\"75.5\" y=\"78.8857\"\u003efrom\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"110.5\" y=\"78.8857\"\u003e\u0026#160;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"60\" x=\"114.5\" y=\"78.8857\"\u003eexternal\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"63\" x=\"17\" y=\"95.1826\"\u003ePostgres\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"80\" y=\"95.1826\"\u003e\u0026#160;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"66\" x=\"84\" y=\"95.1826\"\u003edatabase\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"150\" y=\"95.1826\"\u003e\u0026#160;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"32\" x=\"154\" y=\"95.1826\"\u003eover\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"28\" x=\"87.5\" y=\"111.4795\"\u003eTCP\u003c/text\u003e\u003c/g\u003e\u003c!--entity 1--\u003e\u003cg id=\"elem_1\"\u003e\u003cpath d=\"M314,45 C314,35 367.5,35 367.5,35 C367.5,35 421,35 421,45 L421,86.5938 C421,96.5938 367.5,96.5938 367.5,96.5938 C367.5,96.5938 314,96.5938 314,86.5938 L314,45 \" fill=\"#B3B3B3\" style=\"stroke:#A6A6A6;stroke-width:0.5;\"/\u003e\u003cpath d=\"M314,45 C314,55 367.5,55 367.5,55 C367.5,55 421,55 421,45 \" fill=\"none\" style=\"stroke:#A6A6A6;stroke-width:0.5;\"/\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"16\" font-weight=\"bold\" lengthAdjust=\"spacing\" textLength=\"87\" x=\"324\" y=\"73.8516\"\u003eDatabase\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"12\" font-style=\"italic\" lengthAdjust=\"spacing\" textLength=\"61\" x=\"337\" y=\"88.7637\"\u003e[Postgres]\u003c/text\u003e\u003c/g\u003e\u003c!--link 0 to 1--\u003e\u003cg id=\"link_0_1\"\u003e\u003cpath d=\"M196.031,66 C232.511,66 273.216,66 305.809,66 \" fill=\"none\" id=\"0-to-1\" style=\"stroke:#666666;stroke-width:1.0;\"/\u003e\u003cpolygon fill=\"#666666\" points=\"313.913,66,305.913,63,305.913,69,313.913,66\" style=\"stroke:#666666;stroke-width:1.0;\"/\u003e\u003ctext fill=\"#666666\" font-family=\"sans-serif\" font-size=\"12\" font-weight=\"bold\" lengthAdjust=\"spacing\" textLength=\"42\" x=\"214.5\" y=\"32.1387\"\u003ereads\u003c/text\u003e\u003ctext fill=\"#666666\" font-family=\"sans-serif\" font-size=\"12\" font-weight=\"bold\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"256.5\" y=\"32.1387\"\u003e\u0026#160;\u003c/text\u003e\u003ctext fill=\"#666666\" font-family=\"sans-serif\" font-size=\"12\" font-weight=\"bold\" lengthAdjust=\"spacing\" textLength=\"35\" x=\"260.5\" y=\"32.1387\"\u003efrom\u003c/text\u003e\u003ctext fill=\"#666666\" font-family=\"sans-serif\" font-size=\"12\" font-weight=\"bold\" lengthAdjust=\"spacing\" textLength=\"69\" x=\"220.5\" y=\"46.1074\"\u003edatabase\u003c/text\u003e\u003ctext fill=\"#666666\" font-family=\"sans-serif\" font-size=\"12\" font-style=\"italic\" lengthAdjust=\"spacing\" textLength=\"32\" x=\"239\" y=\"60.0762\"\u003e[TCP]\u003c/text\u003e\u003c/g\u003e\u003crect fill=\"none\" height=\"16.2969\" style=\"stroke:none;stroke-width:1.0;\" width=\"164\" x=\"243\" y=\"148.7813\"/\u003e\u003ctext fill=\"#000000\" font-family=\"sans-serif\" font-size=\"14\" font-weight=\"bold\" lengthAdjust=\"spacing\" textLength=\"57\" x=\"243\" y=\"161.7764\"\u003eLegend\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"300\" y=\"161.7764\"\u003e\u0026#160;\u003c/text\u003e\u003crect fill=\"#438DD5\" height=\"16.2969\" style=\"stroke:none;stroke-width:1.0;\" width=\"164\" x=\"243\" y=\"165.0781\"/\u003e\u003ctext fill=\"#3C7FC0\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"8\" x=\"247\" y=\"178.0732\"\u003e\u0026#9647;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"255\" y=\"178.0732\"\u003e\u0026#160;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"69\" x=\"263\" y=\"178.0732\"\u003econtainer\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"336\" y=\"178.0732\"\u003e\u0026#160;\u003c/text\u003e\u003crect fill=\"#B3B3B3\" height=\"16.2969\" style=\"stroke:none;stroke-width:1.0;\" width=\"164\" x=\"243\" y=\"181.375\"/\u003e\u003ctext fill=\"#A6A6A6\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"8\" x=\"247\" y=\"194.3701\"\u003e\u0026#9647;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"255\" y=\"194.3701\"\u003e\u0026#160;\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"136\" x=\"263\" y=\"194.3701\"\u003eexternal_container\u003c/text\u003e\u003ctext fill=\"#FFFFFF\" font-family=\"sans-serif\" font-size=\"14\" lengthAdjust=\"spacing\" textLength=\"4\" x=\"403\" y=\"194.3701\"\u003e\u0026#160;\u003c/text\u003e\u003cline style=\"stroke:none;stroke-width:1.0;\" x1=\"243\" x2=\"407\" y1=\"148.7813\" y2=\"148.7813\"/\u003e\u003cline style=\"stroke:none;stroke-width:1.0;\" x1=\"243\" x2=\"407\" y1=\"165.0781\" y2=\"165.0781\"/\u003e\u003cline style=\"stroke:none;stroke-width:1.0;\" x1=\"243\" x2=\"407\" y1=\"181.375\" y2=\"181.375\"/\u003e\u003cline style=\"stroke:none;stroke-width:1.0;\" x1=\"243\" x2=\"407\" y1=\"197.6719\" y2=\"197.6719\"/\u003e\u003cline style=\"stroke:none;stroke-width:1.0;\" x1=\"243\" x2=\"243\" y1=\"148.7813\" y2=\"197.6719\"/\u003e\u003cline style=\"stroke:none;stroke-width:1.0;\" x1=\"407\" x2=\"407\" y1=\"148.7813\" y2=\"197.6719\"/\u003e\u003ctext fill=\"#888888\" font-family=\"sans-serif\" font-size=\"10\" lengthAdjust=\"spacing\" textLength=\"250\" x=\"87\" y=\"226.9541\"\u003egenerated by diagramastext.dev - 2023-04-10\u003c/text\u003e\u003c!--SRC=[JOtBReCm44Nt-OefKXMG2hHILzq2IXUXHQHLbiZ64sB9sCWUqkJlE_ILUZ6IxvnxvaRRtimAuKWqXQSyz-8Z6pGTPpa7zBspX9QotetvP8IbUJHf86Mqp8l7j5cYztgRZo8GUewwWXj2M_JPnEpgu1ml81gG8q6eG5v0QJ5uyTKvKwRm12dSAjx6wmk_jAvJfTP9jFgJnVTt4ErHmWxz2Nt4lurRPej21JXuDmAxq5jXe7611ey1M2ca20YEE_1MD55oLPQogyuKFx2a_E4MuM-PqHPDrowN5yPV3wb_-BTqz_owxxRLfdefu-GJ]--\u003e\u003c/g\u003e\u003c/svg\u003e" }
      type: object
      required:
        - "svg"
      additionalProperties: false
      properties:
        svg:
          description: "Generated diagram encoded in unicode SVG format."
          type: "string"
    Error:
      type: object
      required:
        - "error"
      additionalProperties: false
      properties:
        error:
          description: "Error message"
          type: "string"
          minLength: 1
    Quotas:
      example: {
        prompt_length_max: 300,
        rate_minute: {
          limit: 3,
          used: 1,
          reset: 1681123672
        },
        rate_day: {
          limit: 20,
          used: 1,
          reset: 1681210072
        }
      }
      type: object
      required:
        - rate_minute
        - rate_day
        - prompt_length_max
      additionalProperties: false
      properties:
        prompt_length_max:
          description: "Maximum prompt length allowed"
          type: "number"
        rate_minute:
          description: "Rate Per Minute limit"
          $ref: "#/components/schemas/QuotasRate"
        rate_day:
          description: "Rate Per Day limit"
          $ref: "#/components/schemas/QuotasRate"
    QuotasRate:
      type: object
      required:
        - limit
        - used
        - reset
      additionalProperties: false
      properties:
        limit:
          description: "Quota's maximal limit"
          type: "integer"
          minimum: 1
        used:
          description: "Quota used since previous reset"
          type: "integer"
          minimum: 0
        reset:
          description: "The timestamp of upcoming quota rest in unix epochs"
          type: "integer"
          minimum: 0